requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss_cpu
streamlit
ipython
python-docx